+++ python/ffsim/variational/ucj_spin_balanced.py
