+++ tests/python/variational/ucj_spin_unbalanced_test.py
@@ -8,7 +8,7 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""Tests for open-shell unitary cluster Jastrow ansatz."""
+"""Tests for spin-unbalanced unitary cluster Jastrow ansatz."""
 
 import itertools
